move the handleSubmit function completely to the cloud functions to allow researchers to record points for feedback coding
